+++ test/Sentry.Tests/ModuleInit.cs
@@ -5,7 +5,6 @@ public static class ModuleInit
     [ModuleInitializer]
     public static void Init()
     {
-        VerifyDiffPlex.Initialize();
         VerifierSettings.IgnoreMembers<SentryException>(_ => _.Module, _ => _.ThreadId);
 
         VerifierSettings.MemberConverter<Breadcrumb, IReadOnlyDictionary<string, string>>(